# docs/swagger.yaml
definitions:
  bo.BaseOrgInfoBo:
    properties:
      orgId:
        type: integer
      orgName:
        type: string
      orgOwnerId:
        type: integer
      outOrgId:
        type: string
      sourceChannel:
        type: string
    type: object
...
  inner_req.AddOrgOutCollaborator:
    properties:
      orgId:
        type: integer
      userId:
        type: integer
    type: object
...
  inner_req.DeptListByIdsInnerReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
      orgId:
        description: OrgId 组织ID
        type: integer
    type: object
  inner_req.GetDeptUserIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.GetLeadersByDeptIdsReq:
    properties:
      deptIds:
        description: 部门id
        items:
          type: integer
        type: array
      orgId:
        description: 组织id
        type: integer
    type: object
...
  inner_req.GetRoleUserIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.GetUserCountByDeptIdsReq:
    properties:
      deptIds:
        description: 部门id
        items:
          type: integer
        type: array
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.GetUserDeptIdsBatchReq:
    properties:
      orgId:
        description: 组织id
        type: integer
      userIds:
        description: 人员id
        items:
          type: integer
        type: array
    type: object
  inner_req.GetUserDeptIdsReq:
    properties:
      orgId:
        description: 组织id
        type: integer
      userId:
        description: 人员id
        type: integer
    type: object
  inner_req.GetUserIdsByDeptIdsReq:
    properties:
      deptIds:
        description: 部门id
        items:
          type: integer
        type: array
      orgId:
        description: 组织id
        type: integer
    type: object
  inner_req.RoleListByIdsInnerReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
      orgId:
        description: OrgId 组织ID
        type: integer
    type: object
...
  inner_resp.DepartmentLeader:
    properties:
      departmentId:
        type: integer
      leaderId:
        type: integer
    type: object
  inner_resp.DeptInfoInnerResp:
    properties:
      code:
        type: string
      id:
        type: integer
      isDelete:
        description: IsDelete
        type: integer
      isHide:
        type: integer
      name:
        type: string
      orgId:
        type: integer
      parentId:
        type: integer
      path:
        type: string
      sort:
        type: integer
      sourceChannel:
        type: string
      status:
        type: integer
    type: object
  inner_resp.GetDeptUserIdsResp:
    properties:
      data:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
    type: object
  inner_resp.GetLeadersByDeptIdsResp:
    properties:
      leaders:
        items:
          $ref: '#/definitions/inner_resp.DepartmentLeader'
        type: array
    type: object
...
  inner_resp.GetRoleUserIdsResp:
    properties:
      data:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
    type: object
  inner_resp.GetUserCountByDeptIdsResp:
    properties:
      userCount:
        additionalProperties:
          type: integer
        type: object
    type: object
  inner_resp.GetUserDeptIdsBatchResp:
    properties:
      data:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
    type: object
  inner_resp.GetUserDeptIdsResp:
    properties:
      deptIds:
        items:
          type: integer
        type: array
    type: object
  inner_resp.GetUserIdsByDeptIdsResp:
    properties:
      userIds:
        items:
          type: integer
        type: array
    type: object
...
  inner_resp.RoleInfoInnerResp:
    properties:
      id:
        type: integer
      name:
        type: string
      orgId:
        type: integer
      roleGroupId:
        type: integer
    type: object
...
info:
  contact: {}
  title: Ying Sou Sou Apis
paths:
...
  /usercenter/inner/api/v1/dept/getAllListByIds:
    post:
      consumes:
      - application/json
      description: 根据部门ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeptListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取所有部门信息列表
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getLeadersByDeptIds:
    post:
      consumes:
      - application/json
      description: 获取部门下的所有leader
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetLeadersByDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inner_resp.GetLeadersByDeptIdsResp'
        "400":
          description: ""
      summary: 获取部门下的所有leader
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getListByIds:
    post:
      consumes:
      - application/json
      description: 根据部门ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.DeptListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.DeptInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取部门信息列表（可用）
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getUserCountByDeptIds:
    post:
      consumes:
      - application/json
      description: 统计部门下的人数
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetUserCountByDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetUserCountByDeptIdsResp'
            type: array
        "400":
          description: ""
      summary: 统计部门下的人数
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getUserDeptIds:
    post:
      consumes:
      - application/json
      description: 获取人员所在的部门id集合
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetUserDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetUserDeptIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取人员所在的部门id集合
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getUserDeptIdsBatch:
    post:
      consumes:
      - application/json
      description: 获取人员所在的部门id集合（批量）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetUserDeptIdsBatchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inner_resp.GetUserDeptIdsBatchResp'
        "400":
          description: ""
      summary: 获取人员所在的部门id集合（批量）
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getUserIds:
    post:
      consumes:
      - application/json
      description: 获取部门用户（部门对应用户数组）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetDeptUserIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetDeptUserIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取部门用户（部门对应用户数组）
      tags:
      - 部门（内部调用）
  /usercenter/inner/api/v1/dept/getUserIdsByDeptIds:
    post:
      consumes:
      - application/json
      description: 获取部门下的所有人员id
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetUserIdsByDeptIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetUserIdsByDeptIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取部门下的所有人员id
      tags:
      - 部门（内部调用）
...
  /usercenter/inner/api/v1/org/add-out-collaborator:
    post:
      consumes:
      - application/json
      description: 新增外部协作人
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.AddOrgOutCollaborator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bo.BaseOrgInfoBo'
        "400":
          description: ""
      summary: 新增外部协作人
      tags:
      - 企业（内部调用）
  /usercenter/inner/api/v1/org/info:
    get:
      consumes:
      - application/json
      description: 获取企业基础信息
      parameters:
      - description: 组织id
        in: query
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bo.BaseOrgInfoBo'
        "400":
          description: ""
      summary: 获取企业基础信息
      tags:
      - 企业（内部调用）
  /usercenter/inner/api/v1/role/getAllListByIds:
    post:
      consumes:
      - application/json
      description: 根据角色ID列表获取角色信息列表（不分状态和是否删除）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.RoleListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.RoleInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取角色信息列表 （不分状态和是否删除）
      tags:
      - 角色（内部调用）
  /usercenter/inner/api/v1/role/getListByIds:
    post:
      consumes:
      - application/json
      description: 根据角色ID列表获取角色信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.RoleListByIdsInnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.RoleInfoInnerResp'
            type: array
        "400":
          description: ""
      summary: 获取角色信息列表
      tags:
      - 角色（内部调用）
  /usercenter/inner/api/v1/role/getUserIds:
    post:
      consumes:
      - application/json
      description: 获取角色用户（角色对应用户数组）
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inner_req.GetRoleUserIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inner_resp.GetRoleUserIdsResp'
            type: array
        "400":
          description: ""
      summary: 获取角色用户（角色对应用户数组）
      tags:
      - 角色（内部调用）
...
swagger: "2.0"